WIP: Redo RenderSettings list management.
TODO: move add/delete/rename to manager class.
